% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rando_strat.R
\name{rando}
\alias{rando}
\title{Randomisation Stratifiée par Blocs}
\usage{
rando(
  k,
  seed,
  block_sizes,
  nb_block,
  strat = NULL,
  ratio = NULL,
  arm_label = NULL,
  arm_code = NULL,
  strat_label = NULL,
  strat_code = NULL
)
}
\arguments{
\item{k}{nombre de groupes de traitement}

\item{seed}{graine aléatoire pour reproductibilité}

\item{block_sizes}{vecteur numérique définissant la taille réélle des blocs souhaitée}

\item{nb_block}{vecteur numérique définissant le nombre de blocs pour chaque éléments de block_size}

\item{strat}{nombre de strates}

\item{ratio}{Vecteur numérique définissant le ratio d’allocation dans chaque bloc;
Si NULL, un ratio équilibré 1:1:...:1 est utilisé.}

\item{arm_label}{Vecteur de caractère contenant les noms des groupes de traitement;
Si NULL, "Groupe1", "Groupe2"... sont utilisés.}

\item{arm_code}{Vecteur numérique contenant les numeros des groupes de traitement;
Si NULL, 1, 2, ..., k sont utilisés.}

\item{strat_label}{Vecteur de caractère contenant les noms des strates;
Si NULL, "Strate1", "Strate2"... sont utilisés.}

\item{strat_code}{Vecteur numérique contenant les numeros des strates;
Si NULL, 1, 2, ..., strat sont utilisés.}

\item{n}{nombre total de sujets}
}
\value{
n dataframe contenant:l’identifiant du sujet, le numéro et le libellé du groupe de traitement assigné,
ainsi que le code et le libellé de la strate d'appartenance ,
}
\description{
Génère une liste de randomisation stratifiée en appliquant une
randomisation par blocs dans chaque strate. Les strates sont définies
par l'utilisateur via leur nombre et leurs labels.
}
\examples{
rando_strat(n = 40, k = 2, strat = 3, seed = 42,block_sizes = c(4), nb_block = c(10),
strat_label = c("CentreA", "CentreB", "CentreC"))

}
