% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rando_bloc.R
\name{rando_bloc}
\alias{rando_bloc}
\title{Randomisation par blocs}
\usage{
rando_bloc(
  n,
  k,
  seed,
  block_sizes,
  nb_block,
  ratio = NULL,
  arm_label = NULL,
  arm_code = NULL
)
}
\arguments{
\item{n}{nombre total de sujets}

\item{k}{nombre de groupes de traitement}

\item{seed}{graine aléatoire pour reproductibilité}

\item{block_sizes}{vecteur numérique définissant la taille réélle des blocs souhaitée}

\item{nb_block}{vecteur numérique définissant le nombre de blocs pour chaque éléments de block_size}

\item{ratio}{Vecteur numérique définissant le ratio d’allocation dans chaque bloc;
Si NULL, un ratio équilibré 1:1:...:1 est utilisé.}

\item{arm_label}{Vecteur de caractère contenant les noms des groupes de traitement;
Si NULL, "Groupe1", "Groupe2"... sont utilisés.}

\item{arm_code}{Vecteur numérique contenant les numeros des groupes de traitement;
Si NULL, 1, 2, ..., k sont utilisés.}
}
\value{
Un dataframe contenant:l’identifiant du sujet, le numéro et le libellé du groupe de traitement assigné
}
\description{
Génère une liste de randomisation par blocs en attribuant à chaque sujet
un groupe de traitement selon un ratio donné. Les sujets sont assignés à
des blocs de taille définis par l'utilisateur.
}
\examples{
rando_bloc(n = 60, k = 2, seed = 42, block_sizes = c(6), nb_block = c(10),ratio = c(2, 1),arm_label = c("Traitement", "Placebo"))
}
