% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_mean.R
\name{sample_size_mean}
\alias{sample_size_mean}
\title{Calcul de la taille d'\u00e9chantillon pour la comparaison de deux moyennes}
\usage{
sample_size_mean(
  mu1 = NULL,
  mu2 = NULL,
  sd = NULL,
  sd1 = NULL,
  sd2 = NULL,
  power = 0.8,
  alpha = 0.05,
  kappa = 1,
  missing_rate = 0,
  nsim = 10000,
  alternative = "two.sided",
  choice = c("test_student_equilibre", "test_student_desequilibre",
    "test_welch_equilibre", "test_welch_desequilibre", "test_wilcoxon")
)
}
\arguments{
\item{mu1}{Vecteur num\u00e9rique. Moyenne(s) attendue(s) dans le groupe 1.}

\item{mu2}{Vecteur num\u00e9rique. Moyenne(s) attendue(s) dans le groupe 2.}

\item{sd}{Num\u00e9rique. \u00c9cart-type commun aux deux groupes.
Requis pour les tests de Student (\code{"test_student_equilibre"} et
\code{"test_student_desequilibre"}).}

\item{sd1}{Num\u00e9rique. \u00c9cart-type du groupe 1.
Requis pour les tests de Welch et Wilcoxon.}

\item{sd2}{Num\u00e9rique. \u00c9cart-type du groupe 2.
Requis pour les tests de Welch et Wilcoxon.}

\item{power}{Vecteur num\u00e9rique. Puissance(s) souhait\u00e9e(s), entre 0 et 1.
Par d\u00e9faut \code{0.80}.}

\item{alpha}{Vecteur num\u00e9rique. Niveau(x) de significativit\u00e9, entre 0 et 1.
Par d\u00e9faut \code{0.05}.}

\item{kappa}{Num\u00e9rique. Ratio n2/n1 pour les designs d\u00e9s\u00e9quilibr\u00e9s.
Par d\u00e9faut \code{1} (groupes \u00e9quilibr\u00e9s).}

\item{missing_rate}{Vecteur num\u00e9rique. Taux de donn\u00e9es manquantes attendu,
entre 0 et 1. Par d\u00e9faut \code{0}.}

\item{nsim}{Entier. Nombre de simulations pour le test de Wilcoxon.
Par d\u00e9faut \code{10000}.}

\item{alternative}{Caract\u00e8re. Type de test : \code{"two.sided"} (par d\u00e9faut),
\code{"less"} ou \code{"greater"}.}

\item{choice}{Caract\u00e8re. Test statistique \u00e0 utiliser. Valeurs possibles :
\itemize{
  \item \code{"test_student_equilibre"} : test de Student, groupes \u00e9quilibr\u00e9s
  \item \code{"test_student_desequilibre"} : test de Student, groupes d\u00e9s\u00e9quilibr\u00e9s
  \item \code{"test_welch_equilibre"} : test de Welch, groupes \u00e9quilibr\u00e9s
  \item \code{"test_welch_desequilibre"} : test de Welch, groupes d\u00e9s\u00e9quilibr\u00e9s
  \item \code{"test_wilcoxon"} : test de Wilcoxon-Mann-Whitney (simulation)
}}
}
\value{
Un data.frame contenant une ligne par combinaison de param\u00e8tres,
  avec les colonnes :
  \itemize{
    \item \code{test} : test utilis\u00e9
    \item \code{puissance} : puissance utilis\u00e9e
    \item \code{mu1}, \code{mu2} : moyennes
    \item \code{alpha} : niveau de significativit\u00e9
    \item \code{kappa} : ratio d'allocation
    \item \code{prop_manquant} : taux de donn\u00e9es manquantes
    \item \code{n1}, \code{n2} : effectifs par groupe (arrondis \u00e0 l'entier sup\u00e9rieur)
    \item \code{n_total} : effectif total brut
    \item \code{n_total_ajuste_manquant} : effectif total ajust\u00e9 sur le taux de manquants
  }
}
\description{
Calcule la taille d'\u00e9chantillon n\u00e9cessaire pour comparer deux moyennes
selon diff\u00e9rents tests statistiques (Student, Welch, Wilcoxon).
Supporte les groupes \u00e9quilibr\u00e9s et d\u00e9s\u00e9quilibr\u00e9s, et permet d'ajuster
l'effectif sur un taux de donn\u00e9es manquantes.
}
\details{
La fonction g\u00e9n\u00e8re automatiquement toutes les combinaisons possibles des
vecteurs d'arguments via \code{expand.grid()}, et exclut les combinaisons
o\u00f9 \code{mu1 == mu2} (pas de diff\u00e9rence \u00e0 d\u00e9tecter).
Les effectifs sont syst\u00e9matiquement arrondis \u00e0 l'entier sup\u00e9rieur.
L'ajustement sur les donn\u00e9es manquantes est calcul\u00e9 comme :
\code{ceiling(n_total / (1 - missing_rate) / 2) * 2}.
}
\examples{
\dontrun{
# Student \u00e9quilibr\u00e9 avec plusieurs sc\u00e9narios
sample_size_mean(
  mu1 = c(55, 60),
  mu2 = c(50, 52),
  sd  = 10,
  power = c(0.80, 0.90),
  missing_rate = 0.10,
  choice = "test_student_equilibre"
)

# Welch d\u00e9s\u00e9quilibr\u00e9
sample_size_mean(
  mu1   = 60,
  mu2   = 50,
  sd1   = 10,
  sd2   = 15,
  kappa = 2,
  power = 0.80,
  choice = "test_welch_desequilibre"
)
}

}
