% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_prop.R
\name{sample_size_prop}
\alias{sample_size_prop}
\title{Calcul de la taille d'\u00e9chantillon pour la comparaison de deux proportions}
\usage{
sample_size_prop(
  p1 = NULL,
  p2 = NULL,
  power = 0.8,
  alpha = 0.05,
  kappa = 1,
  p01 = NULL,
  p10 = NULL,
  missing_rate = 0,
  alternative = "two.sided",
  choice = c("khi2_equilibre", "khi2_desequilibre", "fisher", "mcnemar")
)
}
\arguments{
\item{p1}{Vecteur num\u00e9rique. Proportion(s) attendue(s) dans le groupe 1,
entre 0 et 1. Non utilis\u00e9 pour McNemar.}

\item{p2}{Vecteur num\u00e9rique. Proportion(s) attendue(s) dans le groupe 2,
entre 0 et 1. Non utilis\u00e9 pour McNemar.}

\item{power}{Vecteur num\u00e9rique. Puissance(s) souhait\u00e9e(s), entre 0 et 1.
Par d\u00e9faut \code{0.80}.}

\item{alpha}{Vecteur num\u00e9rique. Niveau(x) de significativit\u00e9, entre 0 et 1.
Par d\u00e9faut \code{0.05}.}

\item{kappa}{Num\u00e9rique. Ratio n2/n1 pour les designs d\u00e9s\u00e9quilibr\u00e9s.
Par d\u00e9faut \code{1}.}

\item{p01}{Num\u00e9rique. Pour McNemar uniquement. Proportion de paires
discordantes (groupe 1 positif, groupe 2 n\u00e9gatif).}

\item{p10}{Num\u00e9rique. Pour McNemar uniquement. Proportion de paires
discordantes (groupe 1 n\u00e9gatif, groupe 2 positif).}

\item{missing_rate}{Vecteur num\u00e9rique. Taux de donn\u00e9es manquantes attendu,
entre 0 et 1. Par d\u00e9faut \code{0}.}

\item{alternative}{Caract\u00e8re. Type de test : \code{"two.sided"} (par d\u00e9faut),
\code{"less"} ou \code{"greater"}.}

\item{choice}{Caract\u00e8re. Test statistique \u00e0 utiliser. Valeurs possibles :
\itemize{
  \item \code{"khi2_equilibre"} : test du Khi-2, groupes \u00e9quilibr\u00e9s
  \item \code{"khi2_desequilibre"} : test du Khi-2, groupes d\u00e9s\u00e9quilibr\u00e9s
  \item \code{"fisher"} : test exact de Fisher
  \item \code{"mcnemar"} : test de McNemar pour donn\u00e9es appari\u00e9es
}}
}
\value{
Un data.frame contenant une ligne par combinaison de param\u00e8tres,
  avec les colonnes :
  \itemize{
    \item \code{test} : test utilis\u00e9
    \item \code{puissance} : puissance utilis\u00e9e
    \item \code{p1}, \code{p2} : proportions (sauf McNemar)
    \item \code{kappa} : ratio d'allocation (sauf McNemar)
    \item \code{alpha} : niveau de significativit\u00e9
    \item \code{prop_manquant} : taux de donn\u00e9es manquantes
    \item \code{n1}, \code{n2} : effectifs par groupe (sauf McNemar)
    \item \code{n_total} : effectif total brut
    \item \code{n_total_ajuste_manquant} : effectif total ajust\u00e9 sur le taux de manquants
  }
}
\description{
Calcule la taille d'\u00e9chantillon n\u00e9cessaire pour comparer deux proportions
selon diff\u00e9rents tests statistiques (Khi-2, Fisher, McNemar).
Supporte les groupes \u00e9quilibr\u00e9s et d\u00e9s\u00e9quilibr\u00e9s, et permet d'ajuster
l'effectif sur un taux de donn\u00e9es manquantes.
}
\details{
Pour le test de McNemar, \code{p01} et \code{p10} doivent \u00eatre fournis
\u00e0 la place de \code{p1} et \code{p2}, et leur somme ne doit pas d\u00e9passer 1.
Les effectifs sont syst\u00e9matiquement arrondis \u00e0 l'entier sup\u00e9rieur pair.
}
\examples{
\dontrun{
# Khi-2 \u00e9quilibr\u00e9
sample_size_prop(
  p1 = c(0.25, 0.30),
  p2 = c(0.40, 0.50),
  power = c(0.80, 0.90),
  missing_rate = 0.10,
  choice = "khi2_equilibre"
)

# McNemar
sample_size_prop(
  power = 0.80,
  p01   = 0.20,
  p10   = 0.40,
  choice = "mcnemar"
)
}

}
