% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauvegarde.R
\name{sauvegarde}
\alias{sauvegarde}
\title{Sauvegarder une liste de randomisation}
\usage{
sauvegarde(df, format, nom_etude, type, version)
}
\arguments{
\item{df}{Data.frame à exporter}

\item{format}{Format d'export souhaité. Valeurs acceptées :csv et txt}

\item{nom_fichier}{Nom du fichier sans extension.
Si NULL, un nom est généré
 automatiquement sous la forme "liste_randomisation_AAAA-MM-JJ".}
}
\value{
Aucune valeur retournée
}
\description{
Exporte un data.frame issu d'une fonction de randomisation dans un fichier
au format choisi par l'utilisateur. Le nom du fichier peut être personnalisé ;
l'extension est ajoutée automatiquement selon le format sélectionné.
}
\examples{
# Générer une liste de randomisation
ma_liste <- rando_bloc(n = 40, k = 2, seed = 42,
                       block_sizes = c(4), nb_block = c(10))
# Export TXT avec nom personnalisé
sauvegarder(ma_liste, format = "txt", nom_fichier = "essai_pilote")
}
