% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corresp.R
\name{corresp}
\alias{corresp}
\title{FONCTION corresp()}
\usage{
corresp(mini, maxi, seed, statut, version, boi_label = NULL, chemin = NULL)
}
\arguments{
\item{mini}{Entier. Premier numéro de boîte.}

\item{maxi}{Entier. Dernier numéro de boîte.}

\item{seed}{Entier. Graine aléatoire pour la reproductibilité.}

\item{statut}{Caractère. Satut de la liste générée "FINALE" ou "FICTIVE".}

\item{version}{Caractère. Version de la liste générée. Ex : "V01".}

\item{boi_label}{Caractère. Libellé Boîte de traitement.
Si NULL, "Boîte de traitement n°xxx"}

\item{chemin}{Caractère.Caractère.Chemin vers le répertoire de sortie.
Si NULL, répertoire de travail courant .}
}
\value{
Invisible. Exporte les fichiers et affiche leur chemin.
}
\description{
Génère la liste de correspondance boîtes–traitements selon le circuit
défini dans init_essai(), puis exporte automatiquement deux fichiers :
un fichier de données (.txt pour Ennov et .csv  pour REDCap)
un fichier PDF avec les libellés lisibles
}
\examples{
\dontrun{
corresp(mini = 1, maxi = 50, seed = 42, statut = "FICTIVE", version = "v01",
        boi_label = "Kit")

}
}
