% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_init_essai.R, R/aaa_init_essai.R
\name{init_essai}
\alias{init_essai}
\title{FONCTION init_essai()}
\usage{
init_essai(
  nom_etude,
  circuit,
  k,
  block_sizes,
  nb_block,
  ratio = NULL,
  arm_label = NULL,
  arm_code = NULL,
  strat_vars = NULL
)

init_essai(
  nom_etude,
  circuit,
  k,
  block_sizes,
  nb_block,
  ratio = NULL,
  arm_label = NULL,
  arm_code = NULL,
  strat_vars = NULL
)
}
\arguments{
\item{nom_etude}{Caractère. Nom de l'étude .}

\item{circuit}{Caractère. Circuit de randomisation : "ennov" "redcap".}

\item{k}{Entier. Nombre de groupes de traitement (>= 2).}

\item{block_sizes}{Vecteur numérique. Tailles réelles des blocs.}

\item{nb_block}{Vecteur numérique. Nombre de blocs souhaités pour chaque taille.}

\item{ratio}{Vecteur numérique de longueur k. Ratio d'allocation pour chaque bloc.
Si NULL, un ratio équilibré 1:1:...:1 est utilisé..}

\item{arm_label}{Vecteur de caractères de longueur k. Libellés des groupes de traitement.
Si NULL, "Groupe1", "Groupe2"... sont utilisés.}

\item{arm_code}{Vecteur numérique de longueur k. Codes des groupes de traitement.
Si NULL, 1, 2, ..., k sont utilisés..}

\item{strat_vars}{Liste optionnelle décrivant les variables de stratification.
Chaque élément est une liste avec:
codes (vecteur numérique) et labels (vecteur de caractères).
Exemple : list(sexe = list(codes = c(1,2), labels = c("Homme","Femme"))).}
}
\value{
Invisible. Stocke les paramètres dans l'environnement interne.

Invisible. Stocke les paramètres dans l'environnement interne.
}
\description{
Définit, initialise et stocke les paramètres de l'essai clinique dans l'environnement interne
du package. Doit être appelée avant toute génération de liste.

Définit, initialise et stocke les paramètres de l'essai clinique dans l'environnement interne
du package. Doit être appelée avant toute génération de liste.
}
\examples{
\dontrun{
init_essai(
  nom_etude   = "ESSAI_CLINIQUE",
  circuit     = "ennov",
  k           = 2,
  block_sizes = c(4,6),
  nb_block    = c(10,10),
  arm_label   = c("Traitement", "Placebo"),
  strat_vars  = list(
    sexe   = list(codes = c(1, 2), labels = c("Femme", "Homme")),
    centre = list(codes = c(1, 2), labels = c("Centre1", "Centre2"))
  )
)
}
\dontrun{
init_essai(
  nom_etude   = "ESSAI_CLINIQUE",
  circuit     = "ennov",
  k           = 2,
  block_sizes = c(4,6),
  nb_block    = c(10,10),
  arm_label   = c("Traitement", "Placebo"),
  strat_vars  = list(
    sexe   = list(codes = c(1, 2), labels = c("Femme", "Homme")),
    centre = list(codes = c(1, 2), labels = c("Centre1", "Centre2"))
  )
)
}
}
