% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_init_essai.R
\name{init_essai}
\alias{init_essai}
\title{FONCTION init_essai()}
\usage{
init_essai(
  nom_etude,
  circuit,
  k,
  block_sizes,
  nb_block,
  ratio = NULL,
  arm_label = NULL,
  arm_code = NULL,
  strat_vars = NULL,
  id_etude = NULL,
  libelle_etude = NULL,
  investigateur = NULL,
  methodologiste = NULL,
  biostatisticien = NULL,
  confidentiel = FALSE,
  code_usmr = NULL,
  indice_document = NULL
)
}
\arguments{
\item{nom_etude}{Caract\u00e8re. Nom de l'\u00e9tude (utilis\u00e9 dans les noms de fichiers).}

\item{circuit}{Caract\u00e8re. Circuit de randomisation : \code{"ennov"} ou \code{"redcap"}.}

\item{k}{Entier. Nombre de groupes de traitement (>= 2).}

\item{block_sizes}{Vecteur num\u00e9rique. Tailles r\u00e9elles des blocs.}

\item{nb_block}{Vecteur num\u00e9rique. Nombre de blocs souhait\u00e9s pour chaque taille.}

\item{ratio}{Vecteur num\u00e9rique de longueur k. Ratio d'allocation.
Si NULL, un ratio \u00e9quilibr\u00e9 1:1:...:1 est utilis\u00e9.}

\item{arm_label}{Vecteur de caract\u00e8res de longueur k. Libell\u00e9s des groupes.
Si NULL, "Groupe1", "Groupe2"... sont utilis\u00e9s.}

\item{arm_code}{Vecteur num\u00e9rique de longueur k. Codes des groupes.
Si NULL, 1, 2, ..., k sont utilis\u00e9s.}

\item{strat_vars}{Liste optionnelle d\u00e9crivant les variables de stratification.
Chaque \u00e9l\u00e9ment est une liste avec \code{codes} (vecteur num\u00e9rique) et
\code{labels} (vecteur de caract\u00e8res).
Exemple : \code{list(sexe = list(codes = c(1,2), labels = c("Homme","Femme")))}.}

\item{id_etude}{Caract\u00e8re. Identifiant officiel de l'\u00e9tude. Ex : "CHUBX2024/01".
Par d\u00e9faut NULL.}

\item{libelle_etude}{Caract\u00e8re. Libell\u00e9 complet de l'\u00e9tude. Par d\u00e9faut NULL.}

\item{investigateur}{Caract\u00e8re. Nom de l'investigateur coordinateur.
Par d\u00e9faut NULL.}

\item{methodologiste}{Caract\u00e8re. Nom du m\u00e9thodologiste. Par d\u00e9faut NULL.}

\item{biostatisticien}{Caract\u00e8re. Nom du biostatisticien. Par d\u00e9faut NULL.}

\item{confidentiel}{Logique. Si TRUE, affiche la mention CONFIDENTIEL sur
la page de garde. Par d\u00e9faut FALSE.}

\item{code_usmr}{Caract\u00e8re. Code du document USMR. Par d\u00e9faut NULL.}

\item{indice_document}{Caract\u00e8re. Indice du document. Par d\u00e9faut NULL.}
}
\value{
Une liste contenant tous les param\u00e8tres valid\u00e9s de l'essai.
  Doit \u00eatre assign\u00e9e : \code{essai <- init_essai(...)}.
}
\description{
D\u00e9finit et valide les param\u00e8tres de l'essai clinique, puis les retourne
sous forme d'un objet liste. Cet objet doit \u00eatre assign\u00e9 et pass\u00e9 en
argument \u00e0 rand() et corresp().
}
\examples{
\dontrun{
essai <- init_essai(
  nom_etude   = "ESSAI_CLINIQUE",
  circuit     = "ennov",
  k           = 2,
  block_sizes = c(4, 6),
  nb_block    = c(10, 10),
  arm_label   = c("Traitement", "Placebo"),
  strat_vars  = list(
    sexe   = list(codes = c(1, 2), labels = c("Femme", "Homme")),
    centre = list(codes = c(1, 2), labels = c("Centre1", "Centre2"))
  )
)
rand(essai, seed = 42, statut = "FICTIVE", version = "v01")
}
}
