% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_init_essai.R
\name{init_essai}
\alias{init_essai}
\title{FONCTION init_essai()}
\usage{
init_essai(
  nom_etude,
  circuit,
  k,
  block_sizes,
  nb_block,
  ratio = NULL,
  arm_label = NULL,
  arm_code = NULL,
  strat_vars = NULL,
  id_etude = NULL,
  libelle_etude = NULL,
  investigateur = NULL,
  methodologiste = NULL,
  biostatisticien = NULL,
  confidentiel = FALSE,
  code_usmr = NULL,
  indice_document = NULL
)
}
\arguments{
\item{nom_etude}{Caractere. Nom de l'etude (utilise dans les noms de fichiers).}

\item{circuit}{Caractere. Circuit de randomisation : \code{"ennov"} ou \code{"redcap"}.}

\item{k}{Entier. Nombre de groupes de traitement (>= 2).}

\item{block_sizes}{Vecteur numerique. Tailles reelles des blocs.}

\item{nb_block}{Vecteur numerique. Nombre de blocs souhaites pour chaque taille.}

\item{ratio}{Vecteur numerique de longueur k. Ratio d'allocation. Si NULL, un ratio equilibre 1:1:...:1 est utilise.}

\item{arm_label}{Vecteur de caracteres de longueur k. Libelles des groupes. Si NULL, "Groupe1", "Groupe2"... sont utilises.}

\item{arm_code}{Vecteur numerique de longueur k. Codes des groupes. Si NULL, 1, 2, ..., k sont utilises.}

\item{strat_vars}{Liste optionnelle decrivant les variables de stratification.
Chaque element est une liste avec \code{codes} (vecteur numerique) et \code{labels} (vecteur de caracteres).
Exemple : \code{list(sexe = list(codes = c(1,2), labels = c("Homme","Femme")))}.}

\item{id_etude}{Caractere. Identifiant officiel de l'etude. Ex : "CHUBX2024/01". Par defaut NULL.}

\item{libelle_etude}{Caractere. Libelle complet de l'etude. Par defaut NULL.}

\item{investigateur}{Caractere. Nom de l'investigateur principal. Par defaut NULL.}

\item{methodologiste}{Caractere. Nom du methodologiste. Par defaut NULL.}

\item{biostatisticien}{Caractere. Nom du biostatisticien. Par defaut NULL.}

\item{confidentiel}{Logique. Si TRUE, affiche la mention CONFIDENTIEL sur la page de garde. Par defaut FALSE.}

\item{code_usmr}{Caractere. Code du document USMR. Par defaut NULL.}

\item{indice_document}{Caractere. Indice du document. Par defaut NULL.}
}
\value{
Une liste contenant tous les parametres valides de l'essai.
  Doit etre assignee : \code{essai <- init_essai(...)}.
}
\description{
Definit et valide les parametres de l'essai clinique, puis les retourne
sous forme d'un objet liste. Cet objet doit etre assigne et passe en
argument a rand() et corresp().
}
\examples{
\dontrun{
essai <- init_essai(
  nom_etude   = "ESSAI_CLINIQUE",
  circuit     = "ennov",
  k           = 2,
  block_sizes = c(4, 6),
  nb_block    = c(10, 10),
  arm_label   = c("Traitement", "Placebo"),
  strat_vars  = list(
    sexe   = list(codes = c(1, 2), labels = c("Femme", "Homme")),
    centre = list(codes = c(1, 2), labels = c("Centre1", "Centre2"))
  )
)
rand(essai, seed = 42, statut = "FICTIVE", version = "v01")
}
}
