% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_init_essai.R
\name{init_essai}
\alias{init_essai}
\title{FONCTION init_essai()}
\usage{
init_essai(
  nom_etude,
  circuit,
  k,
  block_sizes,
  nb_block,
  ratio = NULL,
  arm_label = NULL,
  arm_code = NULL,
  strat_vars = NULL,
  id_etude = NULL,
  libelle_etude = NULL,
  investigateur = NULL,
  methodologiste = NULL,
  biostatisticien = NULL,
  confidentiel = FALSE,
  code_usmr = NULL,
  indice_document = NULL
)
}
\arguments{
\item{nom_etude}{Caractère. Nom de l'étude .}

\item{circuit}{Caractère. Circuit de randomisation : "ennov" "redcap".}

\item{k}{Entier. Nombre de groupes de traitement (>= 2).}

\item{block_sizes}{Vecteur numérique. Tailles réelles des blocs.}

\item{nb_block}{Vecteur numérique. Nombre de blocs souhaités pour chaque taille.}

\item{ratio}{Vecteur numérique de longueur k. Ratio d'allocation pour chaque bloc.
Si NULL, un ratio équilibré 1:1:...:1 est utilisé..}

\item{arm_label}{Vecteur de caractères de longueur k. Libellés des groupes de traitement.
Si NULL, "Groupe1", "Groupe2"... sont utilisés.}

\item{arm_code}{Vecteur numérique de longueur k. Codes des groupes de traitement.
Si NULL, 1, 2, ..., k sont utilisés..}

\item{strat_vars}{Liste optionnelle décrivant les variables de stratification.
Chaque élément est une liste avec:
codes (vecteur numérique) et labels (vecteur de caractères).
Exemple : list(sexe = list(codes = c(1,2), labels = c("Homme","Femme"))).}

\item{id_etude}{Caractère . Identifiant officiel de l'étude.
Ex : "CHUBX2024/01". Par défaut NULL.}

\item{libelle_etude}{Caractère . Libellé complet de l'étude.
Par défaut NULL.}

\item{investigateur}{Caractère. Nom de l'investigateur coordinateur.
Par défaut NULL.}

\item{methodologiste}{Caractère . Nom du méthodologiste.
Par défaut NULL.}

\item{biostatisticien}{Caractère . Nom du biostatisticien.
Par défaut NULL.}

\item{confidentiel}{Logique. Si TRUE, affiche la mention CONFIDENTIEL sur
la page de garde. Par défaut FALSE.}

\item{code_usmr}{Caractère . Code du document USMR.
Par défaut NULL.}

\item{indice_document}{Caractère . Indice du document.
Par défaut NULL.}
}
\value{
Invisible. Stocke les paramètres dans l'environnement interne.
}
\description{
Définit, initialise et stocke les paramètres de l'essai clinique dans l'environnement interne
du package. Doit être appelée avant toute génération de liste.
}
\examples{
\dontrun{
init_essai(
  nom_etude   = "ESSAI_CLINIQUE",
  circuit     = "ennov",
  k           = 2,
  block_sizes = c(4,6),
  nb_block    = c(10,10),
  arm_label   = c("Traitement", "Placebo"),
  strat_vars  = list(
    sexe   = list(codes = c(1, 2), labels = c("Femme", "Homme")),
    centre = list(codes = c(1, 2), labels = c("Centre1", "Centre2"))
  )
)
}
}
