% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{FONCTION rand()}
\usage{
rand(seed, statut, version, chemin = NULL)
}
\arguments{
\item{seed}{Entier. Graine aléatoire pour la reproductibilité.}

\item{statut}{Caractère. Statut de la liste générée "FINALE" ou "FICTIVE".}

\item{version}{Caractère. Version de la liste générée. Ex : "V01".}

\item{chemin}{Caractère.Chemin vers le répertoire de sortie.
Si NULL, répertoire de travail courant .}
}
\value{
Invisible. Exporte les fichiers et affiche leur chemin.
}
\description{
Génère la liste de randomisation selon le circuit défini dans
init_essai(), puis exporte automatiquement deux fichiers :
un fichier de données (.csv pour Ennov, .txt pour REDCap)
un fichier PDF avec les libellés lisibles
}
\examples{
\dontrun{
rand(seed = 42, type = "FICTIVE", version = "V1")
}

}
