% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_phase2.R
\name{sample_size_phase2}
\alias{sample_size_phase2}
\title{FONCTION sample_size_phase2}
\usage{
sample_size_phase2(
  p0,
  p1,
  alpha = 0.05,
  power = 0.8,
  method = c("ahern", "fleming"),
  missing_rate = 0,
  nmax = 200
)
}
\arguments{
\item{p0}{Proportion de succès sous l'hypothèse nulle (taux inacceptable)}

\item{p1}{Proportion de succès sous l'hypothèse alternative (taux attendu)}

\item{alpha}{Niveau de signification (erreur de type I, défaut = 0.05)}

\item{power}{Puissance souhaitée (1 - erreur de type II, défaut = 0.80)}

\item{method}{Méthode de calcul ("ahern" ou "fleming")}

\item{missing_rate}{Proportion attendue de données manquantes (défaut = 0)}

\item{nmax}{Taille maximale testée pour A'Hern (défaut = 200)}
}
\value{
Un data.frame avec les colonnes : methode, p0, p1, alpha, power,
  missing_prop, n_patient, n_ajuste, n_succes.
}
\description{
Calcule la taille d'échantillon nécessaire pour un essai de phase II à un bras
selon les méthodes de A'Hern ou de Fleming, pour tester si la proportion de succès
attendue est suffisamment élevée pour justifier un essai de phase III.
}
\examples{
 \dontrun{
sample_size_phase2(
  p0 = c(0.10,0.20),
  p1 = c(0.30,0.40),
  method = "ahern",
  missing_rate = 0.1,
  nmax = 50)
  }

}
