% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_report.R
\name{ss_report}
\alias{ss_report}
\title{FONCTION ss_report()}
\usage{
ss_report(
  result,
  file = "sample_size_report.docx",
  nom_etude = NULL,
  investigateur = NULL,
  methodologiste = NULL,
  biostatisticien = NULL
)
}
\arguments{
\item{result}{Objet retourné par
\code{ss_mean_sup()},
\code{ss_mean_ni()},
\code{ss_prop_sup()} ou
\code{ss_prop_ni()}.}

\item{file}{Caractère. Nom du fichier Word à créer. Par défaut \code{"sample_size_report.docx"}.}

\item{investigateur}{Caractère. Nom de l'investigateur principal. Par défaut NULL.}

\item{methodologiste}{Caractère. Nom du méthodologiste. Par défaut NULL.}

\item{biostatisticien}{Caractère. Nom du biostatisticien. Par défaut NULL.}

\item{libelle_etude}{Caractère. Libelle complet de l'étude. Par défaut NULL.}
}
\value{
Crée un document Word et retourne (invisiblement) le chemin du fichier.
}
\description{
Genere un rapport Word (.docx) contenant le resultat d'un calcul
du nombre de sujets necessaire realise avec les fonctions du package.

Le document comporte :
\itemize{
  \item un titre ;
  \item les informations générales de l'étude ;
  \item le tableau des résultats.
}
}
\examples{
\dontrun{

res <- ss_mean_sup(
  mu1 = 55,
  mu2 = 50,
  sd = 10
)

ss_report(
  result = res,
  file = "Calcul_effectif.docx",
  nom_etude = "ESSAI ABC",
  investigateur = "Pr A",
  methodologiste = "Dr B",
  biostatisticien = "C D"
)

}

}
